fix: only run yamllint on catalog and not the rest
All the yaml files except for the catalog are automatically
updated from the internal-services repo (and they are auto generated
by a Makefile there). So every time the files are updated,
they violate yamllint rules. It's better to exclude them
altogether and only run yamllint on the catalog subdir
(which is updated manually).

Signed-off-by: Martin Malina <mmalina@redhat.com>