refactor(RHTAPREL-643): change advisory task error handling
This commit changes the script in the create-advisory-task to trap all
exits instead of just error to improve the output returned to the
internal request. Prior to this change, if the git clone failed, it
would return an error about the checkout of the revision failing as it
would not exit early enough.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>