fix: declare param as env var in advisory task
In order to support string values with single quotes in the
advisory_json parameter of the create-advisory-task, the parameter is
declared as an env variable for the task script instead of using the
parameter directly. This commit also removes the set -x added in PR #96
as that was just added for debugging.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>