feat: add sign-image pipeline
This pipeline yaml is a copy of the pipeline we currently
use: hacbs-signing-pipeline

The goal here is to enable retry in the request-signature
task, see here:
https://issues.redhat.com/browse/RELEASE-1194

For that, I needed to update the task reference. But that
broke the task because the current version of the task
does not have the signature_data result.

In order to test this, I am deploying this modified
pipeline under a different name.

Once this is tested, we may actually switch to using this
going forward - we can evaluate this afterwards.

Signed-off-by: Martin Malina <mmalina@redhat.com>