fix(RELEASE-1271): use bundle resolver in signing pipeline
The old bundle reference was deprecated in Tekton 1.15
and removed in 1.16 and app-sre just deployed 1.16
which broke signing in Konflux. Updating to bundle
resolver format should fix it.

Note that this pipeline definition is still not really
used, we use the hacbs-signing-pipeline from a different
repo, but this is a copy we have and it's useful
for testing.

Signed-off-by: Martin Malina <mmalina@redhat.com>