fix: use staging secret for signing pipelines used in stage
We would like to start using prod RADAS for all signing
(see RELEASE-1191 ). As a first step, we changed the secrets
that are deployed in staging app sre cluster:
app-interface!125757

Previouly, we would have two secrets in the stage cluster
 - prod and nonprod. But both would contain the nonprod
client cert. With this change, the prod contains prod cert
and nonprod contains nonprod cert.

When it got merged, we suddenly started using prod client
cert for all signing in stage cluster (e2e) and we thought
it might work, but apparently it doesn't and broke e2e.

So instead of reverting that app interface change,
let's just change the configMaps to use the nonprod secret.

That way we can keep testing how/if we can use the prod certs
to use prod RADAS.

Signed-off-by: Martin Malina <mmalina@redhat.com>