feat(RELEASE-1191): use prod radas for signing
This commit modifies the signing configMaps in two ways. First, it
changes all the UMB variables to be the production values. This makes it
so that we will use prod radas even in stage signing requests.

Secondly, it removes the SSL_ variables from the configMaps. These were
left for backwards compatibility, but now the PYXIS_ and UMB_ variables
take their places.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>