feat(RELEASE-1191): update prod signing configmaps
This commit updates the stable branch similarly to how the main branch
was updated. The SSL_ keys were removed from the signing configMaps as
the PYXIS_ and UMB_ keys take their place now. Also, all signing
configMaps with stage in their name are set to use prod umb (although I
am not sure the stage configMaps in this branch are used at all).

Signed-off-by: Johnny Bieren <jbieren@redhat.com>